move class into arrow file
